+++ litho-processor/src/main/java/com/facebook/litho/specmodels/model/ClassNames.java
@@ -72,6 +72,7 @@ public interface ClassNames {
 
   ClassName COMPONENT_CONTEXT = ClassName.bestGuess("com.facebook.litho.ComponentContext");
   ClassName COMPONENT_LAYOUT = ClassName.bestGuess("com.facebook.litho.ComponentLayout");
+  ClassName COMPONENT_TREE = ClassName.bestGuess("com.facebook.litho.ComponentTree");
 
   ClassName COMPONENT = ClassName.bestGuess("com.facebook.litho.Component");
   ClassName COMPONENT_BUILDER = ClassName.bestGuess("com.facebook.litho.Component.Builder");